Added some new features to further ease playing music.
Added the Apostrophe key to keep track of wich music track you're currently playing. This is mainly useful for skipping purposes, and you want to remember wich track you were on prior to skipping the music.
You can now hold shift and press left and right arrows to quickly go through all of the music tracks in the folder you're currently focused on.
Holding shift and pressing the letter M will anounce the fade timer's value in ms.
Added the ability to adjust the fade time of the music by holding shift and pressing the up or down arrow keys. This change applies to pausing and resuming the track, as well as exiting the program.