Added some more tweaks to the app.
Fixed the music inside of the sbyw folder to include longer versions of the music that was previously there.
Added sound feedback for exiting the app. This was obtained from an old IOS sounds folder.
Pressing the escape key will now fade the volume of the music instead of fading the pitch upon exiting the program.