Added some more small tweaks.
Made it so the skip sounds don't get adjusted with the other effects while the music is paused or resumed.
Made it so the playback sounds don't get adjusted with the other effects while the music is paused or resumed.
Made it so the shuffle sounds don't get adjusted with the other effects while the music is paused or resumed.