Added back some of the features that were removed in earlier versions of the app.
Added the ability to get the current position of a playing track. You can hold alt and press the letter L to obtain this information.
Added the ability to restart the playing track from the beginning. You can hold alt and press the letter R to do this.
Added back the ability to seek forward and backward through a track. You can hold alt and press left or right arrows to go backward or forward in the current track while it's playing. A sound will also play while you're doing this as well.