Made some much needed improvements to the app.
Made it so music pitch will now be enforced to strictly use semitone shifted values instead of actual raw values for all instences.
Removed the ability to adjust the fade timer, along with speaking it's value in ms.
Made it so the music pitch will no longer fade up or down depending on the playback state of the track.
Fixed a bug where changing the music pitch before and after pausing it would cause the pitch fader to not function propperly upon resuming the music.
Implemented a reverb effect into the app, used later on.